#138 [FEAT] log_out_end

+++ app/src/main/java/com/example/charo_android/presentation/ui/setting/SettingMainFragment.kt
@@ -2,21 +2,27 @@ package com.example.charo_android.presentation.ui.setting
 
 import android.content.Intent
 import android.os.Bundle
+import android.util.Log
 import androidx.fragment.app.Fragment
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
 import android.widget.CompoundButton
 import android.widget.Toast
+import androidx.core.app.ActivityCompat
 import androidx.lifecycle.Observer
 import com.example.charo_android.R
 import com.example.charo_android.databinding.FragmentSettingMainBinding
 import com.example.charo_android.presentation.base.BaseFragment
+import com.example.charo_android.presentation.ui.main.MainActivity
 import com.example.charo_android.presentation.ui.setting.viewmodel.SettingViewModel
 import com.example.charo_android.presentation.ui.signin.SocialSignInActivity
 import com.example.charo_android.presentation.ui.signup.SignUpPassWordFragment
+import com.example.charo_android.presentation.util.CustomDialog
 import com.example.charo_android.presentation.util.SharedInformation
 import com.example.charo_android.presentation.util.changeFragment
+import com.google.firebase.auth.ktx.auth
+import com.google.firebase.ktx.Firebase
 import com.kakao.sdk.user.UserApiClient
 import org.koin.androidx.viewmodel.ext.android.sharedViewModel
 
@@ -29,6 +35,10 @@ class SettingMainFragment : BaseFragment<FragmentSettingMainBinding>(R.layout.fr
         clickProfileUpdate()
         changeTabText()
         clickLogOut()
+
+        Log.d("sharedLog", SharedInformation.getSocialId(requireActivity()))
+        Log.d("sharedLog", SharedInformation.getEmail(requireActivity()))
+
         binding.switchAlarm.setOnCheckedChangeListener(object :
             CompoundButton.OnCheckedChangeListener {
             override fun onCheckedChanged(buttonView: CompoundButton?, isChecked: Boolean) {
@@ -62,19 +72,49 @@ class SettingMainFragment : BaseFragment<FragmentSettingMainBinding>(R.layout.fr
     private fun clickLogOut(){
         val socialKeyStorage = SharedInformation
         val socialKey = socialKeyStorage.getSocialId(requireActivity())
+        val dialog = CustomDialog(requireActivity())
         binding.textSettingLogout.setOnClickListener {
-                if(socialKey == "1"){
-                    UserApiClient.instance.unlink { error ->
-                        if (error != null) {
-                            Toast.makeText(requireActivity(), "로그아웃 실패 $error", Toast.LENGTH_SHORT).show()
-                        }else {
+            SharedInformation.removeEmail(requireActivity())
+            dialog.showDialog(R.layout.custom_dialog_log_out)
+            dialog.setOnClickedListener(object: CustomDialog.ButtonClickListener{
+                override fun onClicked(num: Int) {
+                    if (num == 1){
+                        if(socialKey == "1"){
+                            UserApiClient.instance.unlink { error ->
+                                if (error != null) {
+                                    Toast.makeText(requireActivity(), "로그아웃 실패 $error", Toast.LENGTH_SHORT).show()
+                                }else {
+                                    SharedInformation.removeEmail(requireActivity())
+                                    SharedInformation.removeSocialId(requireActivity())
+                                    Toast.makeText(requireActivity(), "로그아웃 성공", Toast.LENGTH_SHORT).show()
+                                    ActivityCompat.finishAffinity(requireActivity())
+                                    val intent = Intent(requireActivity(), SocialSignInActivity::class.java)
+                                    startActivity(intent)
+
+                                }
+
+                            }
+                        } else if (socialKey == "3"){
+                            SharedInformation.removeEmail(requireActivity())
+                            SharedInformation.removeSocialId(requireActivity())
                             Toast.makeText(requireActivity(), "로그아웃 성공", Toast.LENGTH_SHORT).show()
+                            ActivityCompat.finishAffinity(requireActivity())
+                            val intent = Intent(requireActivity(), SocialSignInActivity::class.java)
+                            startActivity(intent)
+
+                        }else{
+                            Firebase.auth.signOut()
+                            SharedInformation.removeEmail(requireActivity())
+                            SharedInformation.removeSocialId(requireActivity())
+                            Toast.makeText(requireActivity(), "로그아웃 성공", Toast.LENGTH_SHORT).show()
+                            ActivityCompat.finishAffinity(requireActivity())
+                            val intent = Intent(requireActivity(), SocialSignInActivity::class.java)
+                            startActivity(intent)
                         }
-                        val intent = Intent(requireActivity(), SocialSignInActivity::class.java)
-                        startActivity(intent.addFlags(Intent.FLAG_ACTIVITY_CLEAR_TOP))
-                        requireActivity().finish()
                     }
                 }
+            })
+
         }
 
     }
